+++ lib/cpp/src/thrift/concurrency/Mutex.cpp
@@ -17,12 +17,8 @@
  * under the License.
  */
 
-#include <thrift/thrift-config.h>
-
 #include <thrift/concurrency/Mutex.h>
-#include <thrift/concurrency/Util.h>
 
-#include <cassert>
 #include <chrono>
 #include <mutex>
